use pojtl.kemdict.com by default; state progress
Use pojtl.kemdict.com by default because it's not that obvious that we
need to start a server. This is easier, since bulk convert is still
fast.

It's not as fast as doing it with a local instance though (seconds
instead of what feels like less than a second), so add some progress
texts to indicate stuff is happening.